+++ doc/modules/preprocessing_targets.rst
@@ -6,75 +6,10 @@
 Transforming the prediction target (``y``)
 ==========================================
 
-Transforming target in regression
----------------------------------
-
-:class:`TransformedTargetRegressor` transforms the targets ``y`` before fitting a
-regression model. The predictions are mapped back to the original space via an
-inverse transform. It takes as an argument the regressor that will be used for
-prediction, and the transformer that will be applied to the target variable::
-
-  >>> import numpy as np
-  >>> from sklearn.datasets import load_boston
-  >>> from sklearn.preprocessing import (TransformedTargetRegressor,
-  ...                                    QuantileTransformer)
-  >>> from sklearn.linear_model import LinearRegression
-  >>> from sklearn.model_selection import train_test_split
-  >>> boston = load_boston()
-  >>> X = boston.data
-  >>> y = boston.target
-  >>> transformer = QuantileTransformer(output_distribution='normal')
-  >>> regressor = LinearRegression()
-  >>> regr = TransformedTargetRegressor(regressor=regressor,
-  ...                                   transformer=transformer)
-  >>> X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=0)
-  >>> regr.fit(X_train, y_train) # doctest: +ELLIPSIS
-  TransformedTargetRegressor(...)
-  >>> print('R2 score: {0:.2f}'.format(regr.score(X_test, y_test)))
-  R2 score: 0.67
-  >>> raw_target_regr = LinearRegression().fit(X_train, y_train)
-  >>> print('R2 score: {0:.2f}'.format(raw_target_regr.score(X_test, y_test)))
-  R2 score: 0.64
-
-For simple transformations, instead of a Transformer object, a pair of
-functions can be passed, defining the transformation and its inverse mapping::
-
-  >>> from __future__ import division
-  >>> def func(x):
-  ...     return np.log(x)
-  >>> def inverse_func(x):
-  ...     return np.exp(x)
-
-Subsequently, the object is created as::
-
-  >>> regr = TransformedTargetRegressor(regressor=regressor,
-  ...                                   func=func,
-  ...                                   inverse_func=inverse_func)
-  >>> regr.fit(X_train, y_train) # doctest: +ELLIPSIS
-  TransformedTargetRegressor(...)
-  >>> print('R2 score: {0:.2f}'.format(regr.score(X_test, y_test)))
-  R2 score: 0.65
-
-By default, the provided functions are checked at each fit to be the inverse of
-each other. However, it is possible to bypass this checking by setting
-``check_inverse`` to ``False``::
-
-  >>> def inverse_func(x):
-  ...     return x
-  >>> regr = TransformedTargetRegressor(regressor=regressor,
-  ...                                   func=func,
-  ...                                   inverse_func=inverse_func,
-  ...                                   check_inverse=False)
-  >>> regr.fit(X_train, y_train) # doctest: +ELLIPSIS
-  TransformedTargetRegressor(...)
-  >>> print('R2 score: {0:.2f}'.format(regr.score(X_test, y_test)))
-  R2 score: -4.50
-
-.. note::
-
-   The transformation can be triggered by setting either ``transformer`` or the
-   pair of functions ``func`` and ``inverse_func``. However, setting both
-   options will raise an error.
+These are transformers that are not intended to be used on features, only on
+supervised learning targets. See also :ref:`transformed_target_regressor` if
+you want to transform the prediction target for learning, but evaluate the
+model in the original (untransformed) space.
 
 Label binarization
 ------------------
